вынес в util функции
